feat: Add new `run` command
With this new command it adds the ability to run custom commands
and/or scripts right from the terminal.  Examples have been added
to the README and likewise example scripts under the new `examples`
directory.

As a quick overview, before executing anything, we inherit an array
of enviroment variables that contain relevent information of the
package update(s), for example

portsync run 'echo $PACKAGE_ORIGIN: Update to $PACKAGE_LATEST'

Will yield something of the likings:

misc/broot: Update to 1.2.0
devel/gh: Update to 1.36.1